Add verify step to PPL test pipeline
