+++ test/ppl/grant/specs/index.js
@@ -43,18 +43,6 @@ describe('PPL Grant', () => {
     assert.ok(browser.isVisible('h1=Licence granted'));
     console.log('Granted licence');
 
-    gotoGranted(browser, process.env.PROJECT_TITLE);
-    const pdf = downloadFile(browser, 'pdf');
-
-    assert.ok(pdf.includes(process.env.PROJECT_TITLE), 'Project title is displayed');
-    assert.ok(pdf.includes('Basic User'), 'Licence holder name is displayed');
-    assert.ok(pdf.includes('University of Croydon'), 'Primary establishment name is displayed');
-
-    assert.ok(pdf.includes('General constraints'), 'General constraints section is displayed');
-    assert.ok(pdf.includes('Standard conditions'), 'Standard conditions section is displayed');
-
-    assert.ok(pdf.includes('Protocol 1 title'), 'First protocol title is displayed');
-    assert.ok(pdf.includes('Protocol 2 title'), 'Second protocol title is displayed');
   });
 
 });
